+++ components/LogoList/styles.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const Boxlogo = styled('div')`
+  align-items: center;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  margin-top: 1.5rem;
+  @media (max-width: 37.5rem) {
+    flex-direction: column;
+    margin-top: 2.5rem;
+  }
+`;
